+++ OwnTracks/OwnTracks/CertificatesTVC.h
@@ -9,6 +9,7 @@
 #import <UIKit/UIKit.h>
 
 @interface CertificatesTVC : UITableViewController
-@property (strong, nonatomic) NSString *selectedFileName;
+@property (strong, nonatomic) NSString *selectedFileNames;
 @property (strong, nonatomic) NSString *fileNameIdentifier;
+@property (strong, nonatomic) NSNumber *multiple;
 @end