+++ OwnTracks/Pods/Target Support Files/ABStaticTableViewController/ABStaticTableViewController-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_ABStaticTableViewController : NSObject
+@end
+@implementation PodsDummy_ABStaticTableViewController
+@end