+++ OwnTracks/Pods/Target Support Files/FontAwesome/FontAwesome-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_FontAwesome : NSObject
+@end
+@implementation PodsDummy_FontAwesome
+@end