+++ OwnTracks/Pods/Target Support Files/FontAwesomeTools/FontAwesomeTools-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_FontAwesomeTools : NSObject
+@end
+@implementation PodsDummy_FontAwesomeTools
+@end