+++ OwnTracks/Pods/Target Support Files/QRCodeReaderViewController/QRCodeReaderViewController-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_QRCodeReaderViewController : NSObject
+@end
+@implementation PodsDummy_QRCodeReaderViewController
+@end