+++ OwnTracks/Pods/Target Support Files/objc-geohash/objc-geohash-dummy.m
@@ -1,0 +1,5 @@
+#import <Foundation/Foundation.h>
+@interface PodsDummy_objc_geohash : NSObject
+@end
+@implementation PodsDummy_objc_geohash
+@end